requirements.txt
beautifulsoup4==4.12.2
black==23.3.0
requests==2.29.0
PyYAML==6.0